vfs: fix linter errors
- Use fmt.Errorf instead of errors.New(fmt.Sprintf(...)) in rc.go
- Remove unnecessary type conversion in cache.go
- Remove unused function clearActiveCache in rc_cache_status_test.go

+++ vfs/vfscache/cache.go
@@ -939,7 +939,7 @@ func (c *Cache) GetAggregateStats() AggregateStats {
 	}
 
 	if nonErrorItems > 0 {
-		stats.AverageCachePercentage = (totalPercentage + nonErrorItems/2) / int64(nonErrorItems)
+		stats.AverageCachePercentage = (totalPercentage + nonErrorItems/2) / nonErrorItems
 	}
 	return stats
 }
